requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.40.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.0
cssselect2==0.7.0
daphne==4.1.2
defusedxml==0.8.0rc2
Django==5.0.7
django-background-tasks==1.2.8
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-templated-mail==1.1.1
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
fonttools==4.54.1
html5lib==1.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.1
kombu==5.4.2
msgpack==1.0.8
oauthlib==3.2.2
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
pydyf==0.11.0
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyphen==0.16.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
redis==5.0.7
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
setuptools==71.1.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tinycss2==1.3.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
xendit-python @ git+https://github.com/xendit/xendit-python.git@ec58a0d6bf9a6eb6c4d4200932a73094c76914b0
zope.interface==6.4.post2
zopfli==0.2.3.post1
